Upgrade Kotlin to 1.5.30
* Several now-obsolete functions got replaced by their new counterparts.
* kotlin-gradle-plugin, android gradle plugin, kotlinx-coroutines
  got upgraded accordingly.

+++ comboctl/src/commonMain/kotlin/info/nightscout/comboctl/base/TransportLayerIO.kt
@@ -1471,7 +1471,7 @@ open class TransportLayerIO(pumpStateStore: PumpStateStore, private val pumpAddr
         for (i in 0 until 13) {
             val pumpIDByte = packet.payload[4 + i]
             if (pumpIDByte == 0.toByte()) break
-            else pumpIDStrBuilder.append(pumpIDByte.toChar())
+            else pumpIDStrBuilder.append(pumpIDByte.toInt().toChar())
         }
         val pumpID = pumpIDStrBuilder.toString()
